Fix TriggerableFragment of FBAds (#2879)
* [sdk] Update flow-bin

We need latest version of flow-bin in order to properly use Flow with React.createContext.

* [sdk][fbads] Fix Flow types in FacebookAds and triggerable views

TriggerableFragment was not working properly, it was replaced with TriggerableView which is a little bit less convenient, but is much more reliable and straightforward to understand and maintain. Also, node handles to important NativeAd views have been moved to state.

* [ncl][test-suite][docs] Remove mentions of TriggerableFragment from docs and example apps, replace with TriggerableView

fbshipit-source-id: e6e9786